requirements.txt
fastapi
uvicorn[standard]
pydantic
pyyaml
pandas
psutil
jinja2
python-dotenv
requests
